+++ HISTORY.rst
@@ -4,6 +4,9 @@ Changelog
 2.1.0 - Unreleased
 ------------------
 
+* Set debug logging for paramiko when using ``--debug`` option.
+  [fschulze]
+
 * Add ``exec`` command which runs a command on an instance using the Paramiko
   connection, which is faster than invoking an additional ssh call.
   [fschulze]